Mapping
Can now display a real-time map of the arena, showing the position and bounding box of the robot. The Map class in mapping.py is derived from controller.Display. To update the map, call .update() in the main loop.

Also fixed error with position and bearing calculation. Note that the simulation environment uses the relatively uncommon NUE coordinate system.