Various fixes related to numpy types
The controller (which we have no control of) does not accept numpy.int64 as an integer type, and .astype(int) converts the type to int64 instead of int. This caused problems when a np.ndarray is directly unpacked as arguments to functions (i.e. .toList() is not called).

Also fixed dimension of results from matrix multiplication, returning [x, y] instead of [[x, y]].

Refactoring the mapping.py